new in-memory file system design
hopefully cleaner and more scalable.  passes usertests.  no disk back-end
yet; mfsload.c wart for loading existing disk format.  various other
improvements along the way, like a real readdir() call, cleaner ls output,
more correct sh redirection behavior, etc.  same commit because they're
unfortunately bound up with the file system changes.

+++ stdinc/sys/types.h
@@ -8,7 +8,7 @@ typedef ptrdiff_t ssize_t;
 typedef size_t off_t;
 
 typedef int dev_t;
-typedef uint32_t ino_t;
+typedef uint64_t ino_t;
 typedef short nlink_t;
 typedef short mode_t;
 typedef int pid_t;
